UI_Gemini_Docker/fastapi/requirements.txt
fastapi
uvicorn
langchain-google-genai
python-dotenv
yt-dlp
langchain
pydantic
langchain-community